% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwnd-package.R
\docType{package}
\name{rpwnd}
\alias{rpwnd}
\alias{rpwnd-package}
\title{The Most Benignly Malicious R Package on the Internet}
\description{
The Most Benignly Malicious R Package on the Internet
}
\author{
Bob Rudis (bob@rud.is)
}
